requirements.txt
Flask==3.0.3
line-bot-sdk==3.11.0
python-dotenv==0.21.1
requests==2.31.0